+++ tests/kube/service.test.ts
@@ -1,52 +1,44 @@
-import Ingress from '../../src/kube/ingress'; 
+import Service from '../../src/kube/service'; 
 import Node from '../../src/kube/node'; 
 import Icinga from '../../src/icinga';
-import {LoggerInstance} from 'winston'; 
+import Logger from '../../src/logger'; 
 import * as JSONStream from 'json-stream';
 const KubeApi = require('kubernetes-client').Client;
 jest.mock('../../src/icinga');
 jest.mock('../../src/kube/node');
 jest.mock('json-stream');
 jest.mock('kubernetes-client');
+jest.mock('../../src/logger');
 
 const template = {
-    "apiVersion": "extensions/v1beta1",
-    "kind": "Ingress",
+    "apiVersion": "v1",
+    "kind": "Service",
     "metadata": {
-        "annotations": {},
         "name": "foo",
         "namespace": "foobar",
+        "annotations": {}
     },
     "spec": {
-        "rules": [
+        "clusterIP": "10.99.24.32",
+        "ports": [
             {
-                "host": "foobar.example.org",
-                "http": {
-                    "paths": [
-                        {
-                            "backend": {
-                                "serviceName": "foo-backend",
-                                "servicePort": 80
-                            }
-                        }
-                    ]
-                }
+                "name": "http",
+                "port": 80,
+                "protocol": "TCP",
+                "targetPort": 80
             },
             {
-                "host": "barfoo.example.org",
-                "http": {
-                    "paths": [
-                        {
-                            "path": "/foo",
-                            "backend": {
-                                "serviceName": "foo-backend",
-                                "servicePort": 80
-                            }
-                        }
-                    ]
-                }
+                "name": "bar",
+                "port": 10000,
+                "protocol": "tcp",
+                "targetPort": 10000
             }
-        ]
+        ],
+        "selector": {
+            "app": "bar"
+        },
+        "sessionAffinity": "None",
+        "type": "ClusterIP"
     },
     "status": {
         "loadBalancer": {}
@@ -59,39 +51,45 @@ beforeEach(() => {
   fixture = JSON.parse(JSON.stringify(template));
 });
 
-describe('kubernetes ingresses', () => {
-  var instance: Ingress;
+describe('kubernetes services', () => {
+  var instance: Service;
 
-  describe('add ingress object with dummy host', () => {
+  describe('add service object with dummy host', () => {
     it('create icinga host object', () => {
-      let instance = new Ingress(LoggerInstance, Node, KubeApi, Icinga);
+      let instance = new Service(Logger, Node, KubeApi, Icinga);
       Icinga.applyHost = jest.fn();
       instance.prepareObject(fixture);  
       const call = Icinga.applyHost.mock.calls[0];
       expect(call[0]).toBe('foo');
-      expect(call[1]).toBe('foo');
+      expect(call[1]).toBe('10.99.24.32');
       expect(call[2].display_name).toBe('foo');
       expect(call[2].check_command).toBe('dummy');
     });
-
+    
     it('create icinga host object with custom definitions', () => {
-      let instance = new Ingress(LoggerInstance, Node, KubeApi, Icinga, JSONStream, {
-        hostDefinition: {
-          'vars.foo': 'bar',
-          'vars.check_command': 'foo'
+      let instance = new Service(Logger, Node, KubeApi, Icinga, JSONStream, {
+        ClusterIP: {
+          applyServices: true,
+          hostDefinition: {
+            'vars.foo': 'bar',
+            'vars.check_command': 'foo'
+          }
         }
       });
-
+      
       Icinga.applyHost = jest.fn();
       instance.prepareObject(fixture);  
+      console.log(Icinga.applyHost.mock.calls);
       const call = Icinga.applyHost.mock.calls[0];
       expect(call[2]['vars.foo']).toBe('bar');
       expect(call[2]['vars.check_command']).toBe('foo');
     });
-
+    
     it('create icinga host object with templates', () => {
-      let instance = new Ingress(LoggerInstance, Node, KubeApi, Icinga, JSONStream, {
-        hostTemplates: ['foo', 'bar']
+      let instance = new Service(Logger, Node, KubeApi, Icinga, JSONStream, {
+        ClusterIP: {
+          hostTemplates: ['foo', 'bar']
+        }      
       });
 
       Icinga.applyHost = jest.fn();
@@ -99,23 +97,23 @@ describe('kubernetes ingresses', () => {
       const call = Icinga.applyHost.mock.calls[0];
       expect(call[3]).toEqual(['foo', 'bar']);
     });
+    
+    it('do not create icinga host object while service is of type NodePort', () => {
+      let instance = new Service(Logger, Node, KubeApi, Icinga, JSONStream);
 
-    it('do not create icinga host object while attachToNodes is enabled', () => {
-      let instance = new Ingress(LoggerInstance, Node, KubeApi, Icinga, JSONStream, {
-        attachToNodes: true
-      });
-
+      fixture.spec.type = 'NodePort';
       Icinga.applyHost = jest.fn();
       instance.prepareObject(fixture);  
       expect(Icinga.applyHost.mock.instances.length).toBe(0);
     });
   });
   
-  describe('add ingress object namespace as service group', () => {
+  describe('add service object namespace as service group', () => {
     it('create service group per default', async () => {
-      let instance = new Ingress(LoggerInstance, Node, KubeApi, Icinga, JSONStream);
+      let instance = new Service(Logger, Node, KubeApi, Icinga, JSONStream);
 
       Icinga.applyHost = jest.fn();
+      Icinga.hasCheckCommand = jest.fn();
       Icinga.applyService = jest.fn();
       Icinga.applyServiceGroup = jest.fn();
       await instance.prepareObject(fixture);  
@@ -124,7 +122,7 @@ describe('kubernetes ingresses', () => {
     });
     
     it('do not create servicegroup if applyServices is disabled', () => {
-      let instance = new Ingress(LoggerInstance, Node, KubeApi, Icinga, JSONStream, {
+      let instance = new Service(Logger, Node, KubeApi, Icinga, JSONStream, {
         applyServices: false
       });
 
@@ -135,32 +133,69 @@ describe('kubernetes ingresses', () => {
     });
   });
 
-  describe('add all ingress object http path rules as service objects', () => {
+  describe('add all service object ports as service objects', () => {
     it('create all service objects', async () => {
-      let instance = new Ingress(LoggerInstance, Node, KubeApi, Icinga, JSONStream);
+      let instance = new Service(Logger, Node, KubeApi, Icinga, JSONStream);
 
       Icinga.applyService = jest.fn();
       await instance.prepareObject(fixture);  
       const calls = Icinga.applyService.mock.calls;
+     console.log(calls);
       expect(Icinga.applyService.mock.instances.length).toBe(2);
       expect(calls[0][0]).toBe('foo');
       expect(calls[1][0]).toBe('foo');
-      expect(calls[0][1]).toBe('foobar.example.org:http');
-      expect(calls[1][1]).toBe('barfoo.example.org:http');
-      expect(calls[0][2]['vars.http_path']).toBe('/');
-      expect(calls[1][2]['vars.http_path']).toBe('/foo');
+      expect(calls[0][1]).toBe('http');
+      expect(calls[1][1]).toBe('bar');
+
+      expect(calls[0][2]['check_command']).toBe('tcp');
+      expect(calls[0][2]['vars.tcp_port']).toBe(80);
+      expect(calls[1][2]['check_command']).toBe('tcp');
+      expect(calls[1][2]['vars.tcp_port']).toBe(10000);
     });
 
     it('create all service objects with custom service definition', async () => {
-      let instance = new Ingress(LoggerInstance, Node, KubeApi, Icinga, JSONStream, {
-        applyServices: true,
-        serviceDefinition: {
-          'check_command': 'tcp',
-          'vars.foo': 'bar'
+      let instance = new Service(Logger, Node, KubeApi, Icinga, JSONStream, {
+        ClusterIP: {
+          applyServices: true,
+          serviceDefinition: {
+            'check_command': 'http',
+            'vars.foo': 'bar'
+          }
         } 
       });
 
+      Icinga.applyHost = jest.fn();
       Icinga.applyService = jest.fn();
+      Icinga.applyServiceGroup = jest.fn();
+      Icinga.hasCheckCommand = jest.fn();
+      Icinga.hasCheckCommand.mockResolvedValue(true);
+
+      await instance.prepareObject(fixture);  
+      const calls = Icinga.applyService.mock.calls;
+      expect(Icinga.applyService.mock.instances.length).toBe(2);
+      expect(calls[0][2].check_command).toBe('http');
+      expect(calls[0][2]['vars.foo']).toBe('bar');
+      expect(calls[1][2].check_command).toBe('http');
+      expect(calls[1][2]['vars.foo']).toBe('bar');
+    });
+
+    it('create all service objects with custom service definition, check_command not found, fallback to protocol', async () => {
+      let instance = new Service(Logger, Node, KubeApi, Icinga, JSONStream, {
+        ClusterIP: {
+          applyServices: true,
+          serviceDefinition: {
+            'check_command': 'http',
+            'vars.foo': 'bar'
+          }
+        } 
+      });
+
+      Icinga.applyHost = jest.fn();
+      Icinga.applyService = jest.fn();
+      Icinga.applyServiceGroup = jest.fn();
+      Icinga.hasCheckCommand = jest.fn();
+      Icinga.hasCheckCommand.mockResolvedValue(false);
+
       await instance.prepareObject(fixture);  
       const calls = Icinga.applyService.mock.calls;
       expect(Icinga.applyService.mock.instances.length).toBe(2);
@@ -171,9 +206,11 @@ describe('kubernetes ingresses', () => {
     });
 
     it('create all service objects with templates', async () => {
-      let instance = new Ingress(LoggerInstance, Node, KubeApi, Icinga, JSONStream, {
-        applyServices: true,
-        serviceTemplates: ['foo', 'bar']
+      let instance = new Service(Logger, Node, KubeApi, Icinga, JSONStream, {
+        ClusterIP: {
+          applyServices: true,
+          serviceTemplates: ['foo', 'bar']
+        }    
       });
 
       Icinga.applyService = jest.fn();
@@ -183,38 +220,24 @@ describe('kubernetes ingresses', () => {
       expect(calls[0][3]).toEqual(['foo', 'bar']);
       expect(calls[1][3]).toEqual(['foo', 'bar']);
     });
-
-    it('create service objects for tls enabled ingresses', async () => {
-      let instance = new Ingress(LoggerInstance, Node, KubeApi, Icinga, JSONStream, {
-        applyServices: true,
-        serviceTemplates: ['foo', 'bar']
-      });
-
-      fixture.spec.tls = {
-        secretName: 'foo'
-      }
-
-      Icinga.applyService = jest.fn();
-      await instance.prepareObject(fixture);  
-      const calls = Icinga.applyService.mock.calls;
-      expect(Icinga.applyService.mock.instances.length).toBe(4);
-      expect(calls[0][1]).toBe('foobar.example.org:http');
-      expect(calls[1][1]).toBe('foobar.example.org:https');
-      expect(calls[1][2]['vars.http_ssl']).toBe(true);
-    });
-
   });
 
   describe('kubernetes annotations', () => {
     it('check_command/templates annotation', async () => {
-      let instance = new Ingress(LoggerInstance, Node, KubeApi, Icinga, JSONStream, {
-        applyServices: true
+      let instance = new Service(Logger, Node, KubeApi, Icinga, JSONStream, {
+        ClusterIP: {
+          applyServices: true
+        }  
       });
 
       fixture.metadata.annotations['kube-icinga/check_command'] = 'bar';
       fixture.metadata.annotations['kube-icinga/templates'] = 'foobar,barfoo';
-
       Icinga.applyService = jest.fn();
+      Icinga.hasCheckCommand = jest.fn();
+      Icinga.hasCheckCommand.mockResolvedValue(true);
+      Icinga.applyHost = jest.fn();
+      Icinga.applyServiceGroup = jest.fn();
+
       await instance.prepareObject(fixture);
       const calls = Icinga.applyService.mock.calls;
       expect(Icinga.applyService.mock.instances.length).toBe(2);
@@ -225,15 +248,20 @@ describe('kubernetes ingresses', () => {
     });
     
     it('use annotation instead global definition', async () => {
-      let instance = new Ingress(LoggerInstance, Node, KubeApi, Icinga, JSONStream, {
-        applyServices: true,
-        serviceDefinition: {
-          check_command: 'foo'
+      let instance = new Service(Logger, Node, KubeApi, Icinga, JSONStream, {
+        ClusterIP: {
+          applyServices: true,
+          serviceDefinition: {
+            check_command: 'foo'
+          }
         }
       });
       fixture.metadata.annotations['kube-icinga/check_command'] = 'bar';
 
       Icinga.applyService = jest.fn();
+      Icinga.hasCheckCommand = jest.fn();
+      Icinga.hasCheckCommand.mockResolvedValue(true);
+
       await instance.prepareObject(fixture);
       const calls = Icinga.applyService.mock.calls;
       expect(Icinga.applyService.mock.instances.length).toBe(2);
@@ -242,15 +270,20 @@ describe('kubernetes ingresses', () => {
     });
 
     it('definiton merge', async () => {
-      let instance = new Ingress(LoggerInstance, Node, KubeApi, Icinga, JSONStream, {
-        applyServices: true,
-        serviceDefinition: {
-          check_command: 'foo'
+      let instance = new Service(Logger, Node, KubeApi, Icinga, JSONStream, {
+        ClusterIP: {
+          applyServices: true,
+          serviceDefinition: {
+            check_command: 'foo'
+          }
         }
       });
       fixture.metadata.annotations['kube-icinga/definition'] = '{"vars.foo": "foobar"}';
 
       Icinga.applyService = jest.fn();
+      Icinga.hasCheckCommand = jest.fn();
+      Icinga.hasCheckCommand.mockResolvedValue(true);
+
       await instance.prepareObject(fixture);
       const calls = Icinga.applyService.mock.calls;
       expect(Icinga.applyService.mock.instances.length).toBe(2);