+++ CHANGELOG.md
@@ -6,6 +6,11 @@
 * [CHANGE] Upgrade to winston ^3.0.0 stable
 * [CHANGE] Log format is now in json and includes a timestamp
 * [FIX] Service protocol is now lower cases which fixes issues like check_command=TCP insteadof check_command=tcp
+* [FIX] Fixes icinga api username is now correctly read from ICINGA_API_USERNAME as documented (instead ICINGA_API_USER)
+* [FEATURE] Add support for service/host definitions in kubernetes annotations #1
+* [BREAKER!] Removed service.portNameAsCommand option, use kubernetes annotations
+* [FEATURE] Discover persistent volumes and create icinga services for those #4
+* [FIX] Fixed typo Definiton => Definition in default config example
 
 
 ## 1.0.1